+++ servant/CHANGELOG.md
@@ -991,7 +991,7 @@ Compatibility with GHC 9.4, see [PR #1592](https://github.com/haskell-servant/se
   ([#345](https://github.com/haskell-servant/servant/pull/345)
   , [#305](https://github.com/haskell-servant/servant/issues/305))
 
-* Default JSON content type change to `application/json;charset=utf-8`.
+* Default JSON content type change to `application/json`.
   ([#263](https://github.com/haskell-servant/servant/issues/263))
   Related browser bugs:
   [Chromium](https://bugs.chromium.org/p/chromium/issues/detail?id=438464) and